requirements.txt
aiocoap
mysql-connector-python
requests